# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AgentAPI with atomsAgent",
    "description": "HTTP API for Claude Code, Goose, Aider, and atomsAgent chat server.\n\nhttps://github.com/coder/agentapi",
    "version": "0.10.0"
  },
  "paths": {
...
    "/atoms/chat/sessions": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "List Chat Sessions",
        "operationId": "list_chat_sessions_atoms_chat_sessions_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "User identifier",
              "title": "User Id"
            },
            "description": "User identifier"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSessionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/atoms/chat/sessions/{session_id}": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Get Chat Session",
        "operationId": "get_chat_session_atoms_chat_sessions__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "User identifier",
              "title": "User Id"
            },
            "description": "User identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSessionDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/atoms/mcp": {
      "get": {
        "tags": [
          "mcp"
        ],
        "summary": "List Mcp Servers",
        "operationId": "list_mcp_servers_atoms_mcp_get",
        "parameters": [
          {
            "name": "organization_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Organization context",
              "title": "Organization Id"
            },
            "description": "Organization context"
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "User context if applicable",
              "title": "User Id"
            },
            "description": "User context if applicable"
          },
          {
            "name": "include_platform",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include platform-level MCPs",
              "default": true,
              "title": "Include Platform"
            },
            "description": "Include platform-level MCPs"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "mcp"
        ],
        "summary": "Create Mcp Server",
        "operationId": "create_mcp_server_atoms_mcp_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MCPCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPConfiguration"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/atoms/mcp/{mcp_id}": {
      "put": {
        "tags": [
          "mcp"
        ],
        "summary": "Update Mcp Server",
        "operationId": "update_mcp_server_atoms_mcp__mcp_id__put",
        "parameters": [
          {
            "name": "mcp_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "MCP configuration identifier",
              "title": "Mcp Id"
            },
            "description": "MCP configuration identifier"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MCPUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPConfiguration"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "mcp"
        ],
        "summary": "Delete Mcp Server",
        "operationId": "delete_mcp_server_atoms_mcp__mcp_id__delete",
        "parameters": [
          {
            "name": "mcp_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Mcp Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatMessageModel": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "message_index": {
            "type": "integer",
            "title": "Message Index"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tokens"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "message_index",
          "role",
          "content",
          "created_at"
        ],
        "title": "ChatMessageModel"
      },
      "ChatSessionDetailResponse": {
        "properties": {
          "session": {
            "$ref": "#/components/schemas/ChatSessionSummary"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessageModel"
            },
            "type": "array",
            "title": "Messages"
          }
        },
        "type": "object",
        "required": [
          "session",
          "messages"
        ],
        "title": "ChatSessionDetailResponse"
      },
      "ChatSessionListResponse": {
        "properties": {
          "sessions": {
            "items": {
              "$ref": "#/components/schemas/ChatSessionSummary"
            },
            "type": "array",
            "title": "Sessions"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          }
        },
        "type": "object",
        "required": [
          "sessions",
          "total",
          "page",
          "page_size",
          "has_more"
        ],
        "title": "ChatSessionListResponse"
      },
      "ChatSessionSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "organization_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "agent_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Type"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "last_message_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Message At"
          },
          "message_count": {
            "type": "integer",
            "title": "Message Count",
            "default": 0
          },
          "tokens_in": {
            "type": "integer",
            "title": "Tokens In",
            "default": 0
          },
          "tokens_out": {
            "type": "integer",
            "title": "Tokens Out",
            "default": 0
          },
          "tokens_total": {
            "type": "integer",
            "title": "Tokens Total",
            "default": 0
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "archived": {
            "type": "boolean",
            "title": "Archived",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "created_at",
          "updated_at"
        ],
        "title": "ChatSessionSummary"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MCPConfiguration": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "const": "http",
            "title": "Type",
            "default": "http"
          },
          "endpoint": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Endpoint"
          },
          "auth_type": {
            "type": "string",
            "enum": [
              "none",
              "bearer",
              "oauth"
            ],
            "title": "Auth Type",
            "default": "none"
          },
          "bearer_token_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bearer Token Id"
          },
          "oauth_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oauth Provider"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "metadata": {
            "$ref": "#/components/schemas/MCPMetadata"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "scope": {
            "$ref": "#/components/schemas/MCPScope"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "endpoint",
          "scope"
        ],
        "title": "MCPConfiguration"
      },
      "MCPCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "const": "http",
            "title": "Type",
            "default": "http"
          },
          "endpoint": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Endpoint"
          },
          "auth_type": {
            "type": "string",
            "enum": [
              "none",
              "bearer",
              "oauth"
            ],
            "title": "Auth Type",
            "default": "none"
          },
          "bearer_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bearer Token"
          },
          "oauth_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oauth Provider"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "metadata": {
            "$ref": "#/components/schemas/MCPMetadata"
          },
          "scope": {
            "$ref": "#/components/schemas/MCPScope"
          }
        },
        "type": "object",
        "required": [
          "name",
          "endpoint",
          "scope"
        ],
        "title": "MCPCreateRequest"
      },
      "MCPListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MCPConfiguration"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "MCPListResponse"
      },
      "MCPMetadata": {
        "properties": {
          "args": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Args"
          },
          "env": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Env"
          }
        },
        "type": "object",
        "title": "MCPMetadata"
      },
      "MCPScope": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "platform",
              "organization",
              "user"
            ],
            "title": "Type"
          },
          "organization_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "MCPScope"
      },
      "MCPUpdateRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "type": {
            "anyOf": [
              {
                "type": "string",
                "const": "http"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "endpoint": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Endpoint"
          },
          "auth_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "none",
                  "bearer",
                  "oauth"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Auth Type"
          },
          "bearer_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bearer Token"
          },
          "oauth_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oauth Provider"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MCPMetadata"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "MCPUpdateRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}